# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ini adalah dokumentasi API Gateway untuk proyek TELIS (Telkom Legal Intelligence System).",
        "title": "TELIS API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "3.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/chat/feedback": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menyimpan rating jempol dan komentar dari user untuk evaluasi RAG (Data Flywheel). UserID diambil secara otomatis dari token JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Kirim Umpan Balik (HITL)",
                "parameters": [
                    {
                        "description": "Payload Feedback",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.submitFeedbackRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Feedback tersimpan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/redlines": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengunggah dokumen source dan target untuk proses komparasi klausul dan redlining.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Redline"
                ],
                "summary": "Unggah Dokumen Redline",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File PDF Kontrak Lama (Source)",
                        "name": "source_file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File PDF Kontrak Baru (Target)",
                        "name": "target_file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Proses komparasi berhasil di-antrekan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/redlines/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil status (PENDING/PROCESSING/COMPLETED/FAILED) dan hasil komparasi kontrak berdasarkan Job ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Redline"
                ],
                "summary": "Ambil Status \u0026 Hasil Redline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Redline Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status \u0026 Hasil Redline",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found (Job ID tidak ada)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "v1.submitFeedbackRequest": {
            "type": "object",
            "required": [
                "message_id",
                "rating"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "message_id": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}